session-1/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
numpy==1.26.4
openai==1.29.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
